fix: memory leak in MerkleTrie
Node's Buffer uses pooled memory chunk of 8KB by default. The memory
pool won't be freed until all Buffers that reference the pool is freed.
TrieNode creates a bunch of temporary Buffers during MerkleTrie.insert.
And it has a long-lived Buffer that stores the key of an entry.

As a result, the short-lived Buffers creates a lot of holes in memory
pools, but pools can’t be freed because part of them are referenced by
long-lived TrieNodes.

The fix is to create new Uint8Array instances (which doesn't use shared
memory pool) when TrieNodes store the keys. This change also improves
the Buffer usage during TrieNode hash calculation.

With this change, average ArrayBuffer usage is down from 1341 bytes to
44 bytes per entry.