fix: Pin Turbo to 1.7.0 to fix Docker builds
Our previous attempts to fix the build in 50cfacd5 and 1c932091 didn't
work because we had a locally cached layer which lead to an older
version of Turbo being used, which didn't suffer from this bug.

The workaround (for now) is to pin to 1.7.0 explicitly when installing
`turbo` in the Docker build step (which previously wasn't using the
version listed in `package.json`).