Allow duplicate `mentionsPositions` (#596)
It is possible for a cast to have multiple mentions with the same
position. Consider the following cast:

    @a@b@c

In this case, we are mentioning the users `a`, `b`, and `c`, without
separating their names with any other text. The resulting
`mentionsPositions` will be `[0, 0, 0]`, since there is no change in
position.

Relax our validations to allow this case.