chore: Remove unspecified/unrecognized enum values from generated types
Currently, our various enums are created to look like the following:

    enum MyEnum {
      MY_ENUM_NONE = 0,
      ...
      UNRECOGNIZED = -1,
    }

The `*_NONE` exists because proto3 has a hard requirement on the first
value being 0.

The `UNRECOGNIZED` value can be disabled with a flag.

We disable `UNRECOGNIZED` since it doesn't serve much practical purpose
for our uses, and we export a number of types from `@farcaster/js` that
explicitly excludes the `*_NONE` value from the various enums we
reference.

This is important because it makes it easier for us to perform type
narrowing in TypeScript, which makes working with the library more
pleasant/convenient since TypeScript's typechecking can do more heavy
lifting.