fix: Reduce memory consumption of replication example (#943)
We had a report of one user on modern MacBook with 16GB of memory who
was running into heap allocation issues. This was likely due to the
excessive number of promises we were creating (as well as buffering all
data for each call).

Change the logic to fetch in batches of 1K records at most. This slows
down the initial sync but should reduce the likelihood that someone will
hit a memory limit.

We also specify a custom limit in the `yarn start` command so that when
we test locally we are using the same limit as everyone else.