Increase FID lock array to 256 elements
Benchmarking with replayed production gossip messages this change
significantly speeds up processing.

~30 seconds => ~10 seconds to replay ~3,000 gossip messages

When measured I got 2,605 FID locks before the change and 470 after.
This would likely not be as big of an issue if the ThreadPool was
larger. Having a massive fid_locks array shouldn't hurt anything or take
up any measurable memory.

I suggest FID_LOCKS_COUNT gets increased when or if the threadpool size
is increased to continue to keep these conflicts less likely.