feat: Assign shards using jump consistent hashing
While modulo-based arithmetic for assigning shards is simple and works
fine, it does not work well when the number of shards changes.

For example, when increasing the number of shards from 3 → 4, ~83% of
FIDs will be assigned to a different shard. With consistent hashing,
only ~25% will be assigned to a different shard.

Jump consistent hashing is a particular approach to consistent hashing
that is stateless (requires no additional tracking of shard
assignments), which is ideal for hubs, since we don't want hubs to have
to manage per-client tracking of shards.