feat: pull missing messages into sync trie via sync health job (#2326)
There are cases where messages exist in the db but are missing from the
trie. This happens primarily around restarts. When the sync health job
sees "message has already been merged" it's an indication that the
message exists in the db but not in the trie.

Deployed to hoyt for testing and it looks good. 

## Merge Checklist

_Choose all relevant options below by adding an `x` now or at any time
before submitting for review_

- [x] PR title adheres to the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/) standard
- [x] PR has a
[changeset](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#35-adding-changesets)
- [x] PR has been tagged with a change label(s) (i.e. documentation,
feature, bugfix, or chore)
- [ ] PR includes
[documentation](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#32-writing-docs)
if necessary.

<!-- start pr-codex -->

---

## PR-Codex overview
This PR adds missing messages to the sync trie via a sync health job. 

### Detailed summary
- Added missing messages to sync trie
- Updated `processSumbitResults` to handle already merged messages
- Added `numAlreadyMerged` counter
- Updated logging for merged messages

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->