Fname register signature fix (#2384)
## Why is this change needed?

There was a bug due to which fname registration was failing on example
package `hello-world`. The PR fixes this.


<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces a new function for creating user name proof claims
and updates the way user name proof signatures are generated and
processed in the `registerFname` function. It also adds a utility
function for converting bytes to hex.

### Detailed summary
- Added `makeUserNameProofClaim` function import.
- Introduced `bytesToHex` function import.
- Updated `registerFname` to use `await signer.signUserNameProofClaim`
with `makeUserNameProofClaim`.
- Changed signature processing to use `bytesToHex` for
`userNameProofSignature.value`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->