chore: Release 1.19 (#2537)
## Why is this change needed?

Release 1.19. This is a no-op release, and likely the final hub release
before snapchain goes live.

## Merge Checklist

_Choose all relevant options below by adding an `x` now or at any time
before submitting for review_

- [x] PR title adheres to the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/) standard
- [x] PR has a
[changeset](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#35-adding-changesets)
- [x] PR has been tagged with a change label(s) (i.e. documentation,
feature, bugfix, or chore)
- [ ] PR includes
[documentation](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#32-writing-docs)
if necessary.


<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the versioning and release notes for the
`hubble` application, including a version bump in `package.json`, an
update to the `FARCASTER_VERSION`, and the addition of a new version in
the `FARCASTER_VERSIONS_SCHEDULE`.

### Detailed summary
- Updated `CHANGELOG.md` for version `1.19.0` release.
- Bumped `version` in `package.json` from `1.18.0` to `1.19.0`.
- Updated `FARCASTER_VERSION` in `hubble.ts` from `2025.1.8` to
`2025.2.19`.
- Added new version `2025.2.19` to `FARCASTER_VERSIONS_SCHEDULE`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->