fix: shuttle should respect validate messages flag everywhere (#2602)
## Why is this change needed?

Shuttle was ignoring the should validate flag when reconciling link
compact state messages

## Merge Checklist

_Choose all relevant options below by adding an `x` now or at any time
before submitting for review_

- [x] PR title adheres to the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/) standard
- [x] PR has a
[changeset](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#35-adding-changesets)
- [x] PR has been tagged with a change label(s) (i.e. documentation,
feature, bugfix, or chore)
- [ ] PR includes
[documentation](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#32-writing-docs)
if necessary.


<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the version of the `@farcaster/shuttle`
package and making a modification in the `hubEventProcessor.ts` file to
include a new parameter in a function call, enhancing the validation of
messages.

### Detailed summary
- Updated `version` in `package.json` from `0.6.17` to `0.6.18`.
- Added a new entry in `CHANGELOG.md` for version `0.6.18`, noting a fix
for respecting validate messages flag.
- Modified the function call in `hubEventProcessor.ts` to include
`shouldValidate` as an additional argument in
`MessageProcessor.deleteDifferenceMessages`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->